fix more warnings

+++ src/lib/jp2/t1/t1_part1/mqc_dec.cpp
@@ -185,7 +185,7 @@ void mqc_init_dec(mqcoder *mqc, uint8_t *bp, uint32_t len,
     mqc_init_dec_common(mqc, bp, len, extra_writable_bytes);
     mqc_setcurctx(mqc, 0);
     mqc->end_of_byte_stream_counter = 0;
-    mqc->c = ((len==0) ? 0xff : *mqc->bp) << 16;
+    mqc->c = (uint32_t)(((len==0) ? 0xff : *mqc->bp) << 16);
     mqc_bytein_dec(mqc);
     mqc->c <<= 7;
     mqc->ct -= 7;
